Conflicts:
	compiler/frontend/serialization/src/org/jetbrains/jet/descriptors/serialization/ProtoBuf.java
	update_dependencies.xml
        KotlinSpellcheckingStrategy.kt

+++ compiler/testData/codegen/box/primitiveTypes/kt3576.kt
@@ -1,6 +1,6 @@
 object TestObject {
-    val testFloat: Float = 0.9999
-    val otherFloat: Float = 1.01
+    val testFloat: Float = 0.9999.toFloat()
+    val otherFloat: Float = 1.01.toFloat()
 }
 
 fun box(): String {
